Refactor csv_harvester to pull out validations into a module, new iiif_json_harvester
