+++ dlme_airflow/harvester/csv.py
@@ -3,18 +3,9 @@ import logging
 import intake
 import pandas as pd
 
+from harvester.validations import check_equality
+
 catalog = intake.open_catalog("catalog.yaml")
-
-
-def check_equality(harvested_df: pd.DataFrame, saved_df: pd.DataFrame):
-    """Checks for DataFrame equality between latest harvested data with
-    persisted DataFrame.
-
-    @param -- harvested_df
-    @param -- saved_df
-    """
-    if not saved_df.equals(harvested_df):
-        logging.error("harvested dataframe does not equal saved dataframe")
 
 
 def csv_harvester(provider: str):
